requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
telethon>=1.34.0
selenium>=4.0.0
python-dotenv>=1.0.0
